+++ src/types/properties.ts
@@ -8,14 +8,17 @@ export interface DynamicDataSource {
   flowEntityCode: string;
   sysField: boolean;
   type: 'flow' | 'context' | 'database';
-  variableType: 'object' | 'objectList' | 'jsonSchema';
+  variableType: ObjectVariableType;
+  fields: Field[];
 }
 
-export type DataSourceResult = {
+export type DesignDataSourceResult = {
   entities: Entity[];
   fields: Field[];
   selectedEntity: Entity;
   selectedEntityCode: string;
 };
 
-export type UseDataSourceApi = (dataSource: DynamicDataSource) => DataSourceResult;
+export type UseDataSourceApi = (dataSource: DynamicDataSource) => DesignDataSourceResult;
+
+export type ObjectVariableType = 'object' | 'objectList' | 'jsonSchema';