+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.12](https://github.com/qcloud-apaas/web-sdk/compare/v0.1.11...v0.1.12) (2021-08-03)
+
+### [0.1.11](https://github.com/qcloud-apaas/web-sdk/compare/v0.1.10...v0.1.11) (2021-08-03)
+
+
+### Features
+
+* **eventbus:** add useEventBus api ([ad5cea8](https://github.com/qcloud-apaas/web-sdk/commit/ad5cea85294fb3ecf4cd17adc1dd004ff0cac79b))
+
+### [0.1.10](https://github.com/qcloud-apaas/web-sdk/compare/v0.1.9...v0.1.10) (2021-08-03)
+
+
+### Features
+
+* **observer:** add useModel api ([cc5a528](https://github.com/qcloud-apaas/web-sdk/commit/cc5a528fa821badca7591e080418493597f5b257))
+
+### [0.1.9](https://github.com/qcloud-apaas/web-sdk/compare/v0.1.8...v0.1.9) (2021-08-02)
+
+
+### Features
+
+* **canvas:** add useCanvas api ([69eba01](https://github.com/qcloud-apaas/web-sdk/commit/69eba018b56f875288bd88ca9979d12e51cf38fe))
+* **observer:** add componentState api ([fbf6b57](https://github.com/qcloud-apaas/web-sdk/commit/fbf6b57d6bf924942a32098f795ddc4e0c4100db))
+
+### [0.1.8](https://github.com/qcloud-apaas/web-sdk/compare/v0.1.7...v0.1.8) (2021-07-30)
+
+### [0.1.7](https://github.com/qcloud-apaas/web-sdk/compare/v0.1.6...v0.1.7) (2021-07-30)
+
+
+### Features
+
+* **datasource:** fetch datasource api ([a68f7e9](https://github.com/qcloud-apaas/web-sdk/commit/a68f7e92fd067e9c66fb531e9c0a5eb631ef293d))
+
 ### [0.1.6](https://github.com/qcloud-apaas/web-sdk/compare/v0.1.5...v0.1.6) (2021-07-30)